Moved error codes to error.rs.
